+++ docs/docs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### Feat
 
+- add Textual UI
+- add storage cache
+- add boto3 client
 - implement initial project skeleton
 - initialize from calcipy template